# Conflicts:
#	include/common.h
#	include/core2/1EC9740.h
#	include/core2/1ECE0B0.h
#	src/overlays/su/mole.c

+++ include/su/baddiesect.h
@@ -2,7 +2,7 @@
 #define __SU_BADDIE_SECT_H__
 
 #include "common.h"
-s32 _subaddiesect_entrypoint_0(Actor*, u8, u32);
+s32 _subaddiesect_entrypoint_0(Actor*, s32, u32);
 s32 _subaddiesect_entrypoint_1(Actor*);
 void _subaddiesect_entrypoint_2(Actor*, s32, s32);
 #endif
